Turbo refactor

+++ Assets/Classes/Race.cs
@@ -4,17 +4,13 @@ using UnityEngine;
 public class Race : ScriptableObject
 {
     public float attack;
-    public float attackRange;
+    public int attackRange;
     public float attackSpeed;
-    [Range(1f, 1.7f)]
-    public float baseAttackSpeed;
+    [Range(1f, 1.7f)] public float baseAttackSpeed;
     public float health;
     public float defense;
     public float counterAttack;
-    [Range(0f, 1f)]
-    public float evasion;
-    [Range(0f, 1f)]
-    public float magicGift;
-    [Range(0f, 1f)]
-    public float physicsGift;
+    [Range(0f, 1f)] public float evasion;
+    [Range(0f, 1f)] public float magicGift;
+    [Range(0f, 1f)] public float physicsGift;
 }
